oak_abi: move tests module to its own file
